folder structure refactor
