+++ src/pages/Clients.js
@@ -2,17 +2,17 @@ import { CSVLink } from 'react-csv'
 import { Link } from 'react-router-dom'
 import { useEffect, useState } from 'react'
 import { MdDownload } from 'react-icons/md'
-import Header from '../parts/header/Header';
+import Header from '../components/header/Header';
 import data from '../helpers/mock-data.json'
 import Pagination from '../helpers/Pagination';
-import SearchBar from '../parts/searchBar/SearchBar';
+import SearchBar from '../components/searchBar/SearchBar';
 import { Table } from 'react-bootstrap';
 import { FaEllipsisV } from "react-icons/fa";
 import { functions, db } from '../helpers/firebase';
 import { httpsCallable } from 'firebase/functions';
 import useAuth from '../contexts/Auth';
 import { getDocs, collection, deleteDoc, doc } from 'firebase/firestore'
-import ClientModal from '../parts/ClientModal'
+import ClientModal from '../components/ClientModal'
 import { Modal } from 'react-bootstrap'
 import { useForm } from '../hooks/useForm';
