+++ src/pages/Comprehensive.js
@@ -1,10 +1,10 @@
 import { Link } from 'react-router-dom'
 import { useEffect, useState } from 'react'
 import data from '../helpers/mock-data.json'
-import Header from '../parts/header/Header';
+import Header from '../components/header/Header';
 import { Table, Alert } from 'react-bootstrap'
 import Pagination from '../helpers/Pagination';
-import SearchBar from '../parts/searchBar/SearchBar'
+import SearchBar from '../components/searchBar/SearchBar'
 import { FaEllipsisV } from "react-icons/fa";
 
 function Comprehensive() {
