+++ src/pages/admin/AddOrganisation.js
@@ -1,7 +1,7 @@
 import { useEffect } from 'react'
 import { Form } from 'react-bootstrap'
 import '../../assets/styles/addClients.css'
-import Header from '../../parts/header/Header'
+import Header from '../../components/header/Header'
 import { authentication, db } from '../../helpers/firebase'
 import { collection, addDoc } from 'firebase/firestore'
 import { useForm } from '../../hooks/useForm'
