+++ src/pages/admin/SideBar.js
@@ -2,7 +2,7 @@ import '../../assets/styles/menu.css'
 import { Link } from 'react-router-dom'
 import { useState, useEffect } from 'react'
 import { MdLogout  } from 'react-icons/md'
-import DefaultAvatar from '../../parts/DefaultAvatar'
+import DefaultAvatar from '../../components/DefaultAvatar'
 import { Badge } from 'react-bootstrap'
 
 export default function SideBar({role, user, displayName}){
