+++ src/pages/agent/AddUsers.js
@@ -4,8 +4,8 @@ import { httpsCallable } from 'firebase/functions'
 import { functions } from '../../helpers/firebase'
 import { useEffect, useState } from 'react'
 import { Form, Row, Col } from 'react-bootstrap'
-import Upload from '../../parts/uploader/Upload'
-import Header from '../../parts/header/Header'
+import Upload from '../../components/uploader/Upload'
+import Header from '../../components/header/Header'
 import { useForm } from '../../hooks/useForm'
 import useAuth from '../../contexts/Auth'
