+++ src/pages/agent/Clients.js
@@ -2,13 +2,13 @@ import { CSVLink } from 'react-csv'
 import { Link } from 'react-router-dom'
 import { useEffect, useState } from 'react'
 import { MdDownload } from 'react-icons/md'
-import Header from '../../parts/header/Header';
+import Header from '../../components/header/Header';
 import data from '../../helpers/mock-data.json'
 import Pagination from '../../helpers/Pagination';
-import SearchBar from '../../parts/searchBar/SearchBar';
+import SearchBar from '../../components/searchBar/SearchBar';
 import { Table, Modal } from "react-bootstrap";
 import { FaEllipsisV } from "react-icons/fa";
-import ClientModal from '../../parts/ClientModal';
+import ClientModal from '../../components/ClientModal';
 import { functions, db } from '../../helpers/firebase';
 import { httpsCallable } from 'firebase/functions';
 import { doc, deleteDoc } from 'firebase/firestore'
