+++ src/pages/superAdmin/Reports.js
@@ -4,8 +4,8 @@ import data from "../../helpers/mock-data.json";
 import Datatable from "../../helpers/DataTable";
 import { MdDownload } from "react-icons/md";
 import { CSVLink } from "react-csv";
-import SearchBar from "../../parts/searchBar/SearchBar";
-import Header from "../../parts/header/Header";
+import SearchBar from "../../components/searchBar/SearchBar";
+import Header from "../../components/header/Header";
 
 function Reports() {
   useEffect(() => {
