+++ src/pages/supervisor/Settings.js
@@ -2,8 +2,8 @@ import { useEffect, useState } from "react";
 import "../../styles/Settings.css";
 import { Form, Row, Col, Alert } from "react-bootstrap";
 import { MdCheckCircle } from "react-icons/md";
-import Header from "../../parts/header/Header";
-import DefaultAvatar from "../../parts/DefaultAvatar";
+import Header from "../../components/header/Header";
+import DefaultAvatar from "../../components/DefaultAvatar";
 import { authentication } from "../../helpers/firebase";
 
 function Settings() {
